+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.6.1
 Release Date: 2014-06-20
 
   * Add (internal) option to keep pop-up window alive after communicating reply to opener.
+  * Fixed issue where we accidentally formatted error messages twice resulting in wrong error messages.
 
 v1.6.0
 -------------